Accept & Run multiple test cases for each request

+++ src/main/java/com/cp/compiler/models/ProcessOutput.java
@@ -16,7 +16,7 @@ public class ProcessOutput {
     @Setter
     private String stdErr;
 
-    private long executionDuration;
+    private int executionDuration;
     
     private int status;
 }
